feat: fix auto-container shutdown (#5)

+++ repos/conductor/src/controller/kube/kube.ts
@@ -19,7 +19,7 @@ import { buildPorts } from './pod/buildPorts'
 import { buildPortMap } from './pod/buildPortMap'
 import { shouldRemove } from './pod/shouldRemove'
 import { shouldHydrate } from './pod/shouldHydrate'
-import { isObj, isEmptyColl } from '@keg-hub/jsutils'
+import { isObj, isEmptyColl, isStr, noOp } from '@keg-hub/jsutils'
 import { buildPodManifest } from './pod/buildPodManifest'
 import { hydrateRoutes } from '../../utils/hydrateRoutes'
 import { buildImgUri } from '../docker/image/buildImgUri'
@@ -41,7 +41,7 @@ export class Kube extends Controller {
   config: TKubeController
   devRouterActive: boolean
   pods: Record<string, TPod>
-  containerMaps: Record<string, TContainerMap>
+  containerMaps: Record<string, TContainerMap>={}
 
   constructor(conductor:Conductor, config:TKubeController){
     super(conductor, config)
@@ -83,7 +83,7 @@ export class Kube extends Controller {
       // Any container not running always just remove it
       // The user can't access it anyways
       if(mapped.state !== EContainerState.Running){
-        this.remove(pod)
+        this.remove(mapped, true)
         return acc
       }
 
@@ -137,7 +137,7 @@ export class Kube extends Controller {
     if(!image) return
 
     const mapped = buildContainerMap(pod, annotations.ports)
-    const doRemove = shouldRemove(pod, image, mapped, annotations)
+    const doRemove = shouldRemove(pod, mapped, annotations)
     if(doRemove) return this.remove(mapped, true)
 
     const doHydrate = shouldHydrate(mapped.state, watchObj.type, pod?.metadata?.deletionTimestamp)
@@ -236,6 +236,20 @@ export class Kube extends Controller {
   }
 
   /**
+   * Gets a single container map from the Docker-Api by reference
+   * @member Kube
+   */
+  get = async (podRef:TPodRef):Promise<TContainerMap> => {
+    const id = isStr(podRef)
+      ? podRef
+      : (podRef as TContainerMap)?.id || (podRef as TPod)?.metadata?.name
+
+    const pod = await this.kubectl.getPod(id) as TPod
+
+    return buildContainerMap(pod, {})
+  }
+
+  /**
    * Removes a container from the runtime cache based on Id
    * @member Kube
    */
@@ -263,14 +277,25 @@ export class Kube extends Controller {
    * Removes a pod from kubernetes, by calling the kubectl API
    * @member Kube
    */
-  remove = async (podRef:TPodRef, isContainerMap:boolean=false) => {
+  remove = async (
+    podRef:TPodRef,
+    isContainerMap:boolean=false,
+    throwOnEmpty:boolean=true
+  ) => {
     if(this.devRouterActive) return
 
     const containerMap = isContainerMap
       ? podRef as TContainerMap
       : this.getContainer(podRef)
 
-    !containerMap && this.notFoundErr({ type: `container`, ref: podRef as string })
+    if(!containerMap){
+      this?.config?.onRemove?.(containerMap)
+
+      throwOnEmpty
+        && this.notFoundErr({ type: `container`, ref: podRef as string })
+
+      return containerMap
+    }
 
     await this.kubectl.deletePod(containerMap.id)
     Logger.info(`Removing container with ID ${containerMap.id}`)
@@ -283,6 +308,7 @@ export class Kube extends Controller {
       }, {})
 
     Logger.success(`Pod ${containerMap.id} removed successfully`)
+    this?.config?.onRemove?.(containerMap)
 
     return containerMap
   }
